+++ Sources/ProjectSpec/SpecValidationError.swift
@@ -20,6 +20,7 @@ public struct SpecValidationError: Error, CustomStringConvertible {
         case invalidSchemeTarget(scheme: String, target: String, action: String)
         case invalidSchemeConfig(scheme: String, config: String)
         case invalidSwiftPackage(name: String, target: String)
+        case invalidPackageDependencyReference(name: String)
         case invalidLocalPackage(String)
         case invalidConfigFile(configFile: String, config: String)
         case invalidBuildSettingConfig(String)
@@ -69,6 +70,8 @@ public struct SpecValidationError: Error, CustomStringConvertible {
                 return "Target \(target.quoted) has an invalid package dependency \(name.quoted)"
             case let .invalidLocalPackage(path):
                 return "Invalid local package \(path.quoted)"
+            case let .invalidPackageDependencyReference(name):
+                return "Package reference \(name) must be specified as package dependency, not target"
             case let .missingConfigForTargetScheme(target, configType):
                 return "Target \(target.quoted) is missing a config of type \(configType.rawValue) to generate its scheme"
             case let .missingDefaultConfig(name):