+++ Tests/XcodeGenKitTests/SchemeGeneratorTests.swift
@@ -54,8 +54,8 @@ class SchemeGeneratorTests: XCTestCase {
                     build: Scheme.Build(targets: [buildTarget], preActions: [preAction]),
                     run: Scheme.Run(config: "Debug", askForAppToLaunch: true, launchAutomaticallySubstyle: "2", simulateLocation: simulateLocation, storeKitConfiguration: storeKitConfiguration, customLLDBInit: "/sample/.lldbinit"),
                     test: Scheme.Test(config: "Debug", targets: [
-                        Scheme.Test.TestTarget(targetReference: TargetReference(framework.name), location: "test.gpx"),
-                        Scheme.Test.TestTarget(targetReference: TargetReference(framework.name), location: "New York, NY, USA")
+                        Scheme.Test.TestTarget(targetReference: TestableTargetReference(framework.name), location: "test.gpx"),
+                        Scheme.Test.TestTarget(targetReference: TestableTargetReference(framework.name), location: "New York, NY, USA")
                     ], customLLDBInit: "/test/.lldbinit"),
                     profile: Scheme.Profile(config: "Release", askForAppToLaunch: true)
                 )
@@ -395,7 +395,8 @@ class SchemeGeneratorTests: XCTestCase {
                         gatherCoverageData: true,
                         coverageTargets: [
                             "TestProject/ExternalTarget",
-                            TargetReference(framework.name),
+                            TestableTargetReference(framework.name),
+                            TestableTargetReference(name: "XcodeGenKitTests", location: .package("XcodeGen"))
                         ]
                     )
                 )
@@ -403,6 +404,7 @@ class SchemeGeneratorTests: XCTestCase {
                     name: "test",
                     targets: [framework],
                     schemes: [scheme],
+                    packages: ["XcodeGen": .local(path: "../")],
                     projectReferences: [
                         ProjectReference(name: "TestProject", path: externalProject.string),
                     ]
@@ -410,7 +412,7 @@ class SchemeGeneratorTests: XCTestCase {
                 let xcodeProject = try project.generateXcodeProject()
                 let xcscheme = try unwrap(xcodeProject.sharedData?.schemes.first)
                 try expect(xcscheme.testAction?.codeCoverageEnabled) == true
-                try expect(xcscheme.testAction?.codeCoverageTargets.count) == 2
+                try expect(xcscheme.testAction?.codeCoverageTargets.count) == 3
                 let buildableReference = xcscheme.testAction?.codeCoverageTargets.first
                 try expect(buildableReference?.blueprintName) == "ExternalTarget"
                 try expect(buildableReference?.referencedContainer) == "container:\(externalProject.string)"
@@ -474,6 +476,33 @@ class SchemeGeneratorTests: XCTestCase {
 
                 let xcscheme = try unwrap(xcodeProject.sharedData?.schemes.first)
                 try expect(xcscheme.launchAction?.macroExpansion?.buildableName) == "MyApp.app"
+            }
+            
+            $0.it("generates scheme with test target of local swift package") {
+                let targetScheme = TargetScheme(
+                    testTargets: [Scheme.Test.TestTarget(targetReference: TestableTargetReference(name: "XcodeGenKitTests", location: .package("XcodeGen")))])
+                let app = Target(
+                    name: "MyApp",
+                    type: .application,
+                    platform: .iOS,
+                    dependencies: [
+                        Dependency(type: .package(product: nil), reference: "XcodeGen")
+                    ],
+                    scheme: targetScheme
+                )
+                let project = Project(
+                    name: "ios_test",
+                    targets: [app],
+                    packages: ["XcodeGen": .local(path: "../")]
+                )
+                let xcodeProject = try project.generateXcodeProject()
+                let xcscheme = try unwrap(xcodeProject.sharedData?.schemes.first)
+                let buildableReference = try unwrap(xcscheme.testAction?.testables.first?.buildableReference)
+
+                try expect(buildableReference.blueprintIdentifier) == "XcodeGenKitTests"
+                try expect(buildableReference.blueprintName) == "XcodeGenKitTests"
+                try expect(buildableReference.buildableName) == "XcodeGenKitTests"
+                try expect(buildableReference.referencedContainer) == "container:../"
             }
 
             $0.it("generates scheme capturing screenshots automatically and deleting on success") {